+++ packages/typescript-estree/src/tsconfig-parser.ts
@@ -1,8 +1,6 @@
-'use strict';
-
 import path from 'path';
 import ts from 'typescript';
-import { Extra } from './temp-types-based-on-js-source';
+import { Extra } from './parser-options';
 
 //------------------------------------------------------------------------------
 // Environment calculation
@@ -10,16 +8,14 @@ import { Extra } from './temp-types-based-on-js-source';
 
 /**
  * Default compiler options for program generation from single root file
- * @type {ts.CompilerOptions}
  */
 const defaultCompilerOptions: ts.CompilerOptions = {
   allowNonTsExtensions: true,
-  allowJs: true
+  allowJs: true,
 };
 
 /**
  * Maps tsconfig paths to their corresponding file contents and resulting watches
- * @type {Map<string, ts.WatchOfConfigFile<ts.SemanticDiagnosticsBuilderProgram>>}
  */
 const knownWatchProgramMap = new Map<
   string,
@@ -29,47 +25,51 @@ const knownWatchProgramMap = new Map<
 /**
  * Maps file paths to their set of corresponding watch callbacks
  * There may be more than one per file if a file is shared between projects
- * @type {Map<string, ts.FileWatcherCallback>}
  */
 const watchCallbackTrackingMap = new Map<string, ts.FileWatcherCallback>();
 
+const parsedFilesSeen = new Set<string>();
+
 /**
  * Holds information about the file currently being linted
- * @type {{code: string, filePath: string}}
  */
 const currentLintOperationState = {
   code: '',
-  filePath: ''
+  filePath: '',
 };
 
 /**
  * Appropriately report issues found when reading a config file
- * @param {ts.Diagnostic} diagnostic The diagnostic raised when creating a program
- * @returns {void}
+ * @param diagnostic The diagnostic raised when creating a program
  */
 function diagnosticReporter(diagnostic: ts.Diagnostic): void {
   throw new Error(
-    ts.flattenDiagnosticMessageText(diagnostic.messageText, ts.sys.newLine)
+    ts.flattenDiagnosticMessageText(diagnostic.messageText, ts.sys.newLine),
   );
 }
 
 const noopFileWatcher = { close: () => {} };
 
+function getTsconfigPath(tsconfigPath: string, extra: Extra): string {
+  return path.isAbsolute(tsconfigPath)
+    ? tsconfigPath
+    : path.join(extra.tsconfigRootDir || process.cwd(), tsconfigPath);
+}
+
 /**
  * Calculate project environments using options provided by consumer and paths from config
- * @param {string} code The code being linted
- * @param {string} filePath The path of the file being parsed
- * @param {string} extra.tsconfigRootDir The root directory for relative tsconfig paths
- * @param {string[]} extra.project Provided tsconfig paths
- * @returns {ts.Program[]} The programs corresponding to the supplied tsconfig paths
+ * @param code The code being linted
+ * @param filePath The path of the file being parsed
+ * @param extra.tsconfigRootDir The root directory for relative tsconfig paths
+ * @param extra.project Provided tsconfig paths
+ * @returns The programs corresponding to the supplied tsconfig paths
  */
 export function calculateProjectParserOptions(
   code: string,
   filePath: string,
-  extra: Extra
+  extra: Extra,
 ): ts.Program[] {
   const results = [];
-  const tsconfigRootDir = extra.tsconfigRootDir;
 
   // preserve reference to code and file being linted
   currentLintOperationState.code = code;
@@ -78,15 +78,12 @@ export function calculateProjectParserOptions(
   // Update file version if necessary
   // TODO: only update when necessary, currently marks as changed on every lint
   const watchCallback = watchCallbackTrackingMap.get(filePath);
-  if (typeof watchCallback !== 'undefined') {
+  if (parsedFilesSeen.has(filePath) && typeof watchCallback !== 'undefined') {
     watchCallback(filePath, ts.FileWatcherEventKind.Changed);
   }
 
-  for (let tsconfigPath of extra.projects) {
-    // if absolute paths aren't provided, make relative to tsconfigRootDir
-    if (!path.isAbsolute(tsconfigPath)) {
-      tsconfigPath = path.join(tsconfigRootDir, tsconfigPath);
-    }
+  for (let rawTsconfigPath of extra.projects) {
+    const tsconfigPath = getTsconfigPath(rawTsconfigPath, extra);
 
     const existingWatch = knownWatchProgramMap.get(tsconfigPath);
 
@@ -103,7 +100,7 @@ export function calculateProjectParserOptions(
       ts.sys,
       ts.createSemanticDiagnosticsBuilderProgram,
       diagnosticReporter,
-      /*reportWatchStatus*/ () => {}
+      /*reportWatchStatus*/ () => {},
     );
 
     // ensure readFile reads the code being linted instead of the copy on disk
@@ -124,7 +121,8 @@ export function calculateProjectParserOptions(
         .getConfigFileParsingDiagnostics()
         .filter(
           diag =>
-            diag.category === ts.DiagnosticCategory.Error && diag.code !== 18003
+            diag.category === ts.DiagnosticCategory.Error &&
+            diag.code !== 18003,
         );
       if (configFileDiagnostics.length > 0) {
         diagnosticReporter(configFileDiagnostics[0]);
@@ -138,7 +136,7 @@ export function calculateProjectParserOptions(
       return {
         close: () => {
           watchCallbackTrackingMap.delete(normalizedFileName);
-        }
+        },
       };
     };
 
@@ -149,7 +147,7 @@ export function calculateProjectParserOptions(
     const oldOnDirectoryStructureHostCreate = (watchCompilerHost as any)
       .onCachedDirectoryStructureHostCreate;
     (watchCompilerHost as any).onCachedDirectoryStructureHostCreate = (
-      host: any
+      host: any,
     ) => {
       const oldReadDirectory = host.readDirectory;
       host.readDirectory = (
@@ -157,7 +155,7 @@ export function calculateProjectParserOptions(
         extensions?: ReadonlyArray<string>,
         exclude?: ReadonlyArray<string>,
         include?: ReadonlyArray<string>,
-        depth?: number
+        depth?: number,
       ) =>
         oldReadDirectory(
           path,
@@ -166,7 +164,7 @@ export function calculateProjectParserOptions(
             : extensions.concat(extra.extraFileExtensions),
           exclude,
           include,
-          depth
+          depth,
         );
       oldOnDirectoryStructureHostCreate(host);
     };
@@ -180,6 +178,7 @@ export function calculateProjectParserOptions(
     results.push(program);
   }
 
+  parsedFilesSeen.add(filePath);
   return results;
 }
 
@@ -187,26 +186,21 @@ export function calculateProjectParserOptions(
  * Create program from single root file. Requires a single tsconfig to be specified.
  * @param code The code being linted
  * @param filePath The file being linted
- * @param {string} extra.tsconfigRootDir The root directory for relative tsconfig paths
- * @param {string[]} extra.project Provided tsconfig paths
- * @returns {ts.Program} The program containing just the file being linted and associated library files
+ * @param extra.tsconfigRootDir The root directory for relative tsconfig paths
+ * @param extra.project Provided tsconfig paths
+ * @returns The program containing just the file being linted and associated library files
  */
 export function createProgram(code: string, filePath: string, extra: Extra) {
   if (!extra.projects || extra.projects.length !== 1) {
     return undefined;
   }
 
-  let tsconfigPath = extra.projects[0];
-
-  // if absolute paths aren't provided, make relative to tsconfigRootDir
-  if (!path.isAbsolute(tsconfigPath)) {
-    tsconfigPath = path.join(extra.tsconfigRootDir, tsconfigPath);
-  }
+  const tsconfigPath = getTsconfigPath(extra.projects[0], extra);
 
   const commandLine = ts.getParsedCommandLineOfConfigFile(
     tsconfigPath,
     defaultCompilerOptions,
-    { ...ts.sys, onUnRecoverableConfigFileDiagnostic: () => {} }
+    { ...ts.sys, onUnRecoverableConfigFileDiagnostic: () => {} },
   );
 
   if (!commandLine) {